+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 CHANGELOG
 =========
 
+1.1 (dev)
+=========
+
+ * added --upgrade to init to perform a configuration upgrade
+ * changed configuration format from INI to YAML (~/.gondor can remain in INI format)
+ * added --no-on-deploy to deploy command which will prevent running of
+   on_deploy commands
+ * improved run to use the new interactive process on Gondor
+
 1.0.6
 =====